[css-layout-api] Adds ability to specify percentage{Inline,Block}Size.
This allows developers to specify the size which percentages in the
block (wrt child writing mode) direction should be resolved against.
Currently inline percentages still resolve against the available
inline-size.

This adds a "OverrideContainingBlockPercentageResolutionLogicalHeight"
field on layout box, which both the logic for
LayoutBox::ComputePercenatgeLogicalHeight and
LayoutBox::ComputeReplacedLogicalHeight use.

They follow similar logic to LayoutGrid, in that in quirks mode,
percentages are treated as indefinite (doesn't "walk-up" the tree).

Change-Id: I256381687d2eed7693a908074f7f5032d5963e2b